Fix issue with relative `dir` values causing invalid in-package calculation
It was found in a reproducer in #1056 that the in-package calculation was failing in certain scenarios. I found in the reproducer that the issue came from the fact that the path describing the outPkgPath was unclean:

```
2025-05-30T16:03:23.240705000-05:00 DBG output package detected to not be in-package of original package file=mock_itf_test.go out-pkg-name=mockerybug outPkgFSPath=/tmp/mockerybug/. outPkgPath=githbub.com/kgadams/mockerybug src-pkg-name=mockerybug srcPkgFSPath=/tmp/mockerybug version=v0.0.0-dev
```

You can see it calculated the output package path to be /tmp/mockerybug/. when it should be /tmp/mockerybug.

As an additional piece of information, this bug is hit only when dir is set to . or ./. We don't encounter this problem if dir is a relative path to a subdirectory.

Fixes #1056.