Fix issue with type alias to interface not being generated
It was found that mockery v3 was not generating mocks for type aliases. This appears
to have been an oversight where type alias support was added in v2 but not in v3.
We simply add similar logic into v3 to allow mockery to consider type aliases declarations
to be acceptable AST nodes in addition to updating its type parsing logic to
explicitly understand how to deal with aliases.

Fixes #1073

Bump version